+++ src/Utilities/Asset.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Grzechu\Utilities;
+namespace RockujemyWpExt\Utilities;
 
 class Asset
 {